Moves FormManagerList and DataManagerList to fragmetns package
